Fixed typo and small flaws
